Move ItemsTable and related components to frontend/src/metabase/components/ItemsTable
