+++ frontend/src/metabase/components/ItemsTable/index.tsx
@@ -1,0 +1,1 @@
+export { ItemsTable } from "./ItemsTable";
