+++ frontend/src/metabase/containers/dnd/ItemsDragLayer.jsx
@@ -3,9 +3,9 @@ import { Component } from "react";
 import { DragLayer } from "react-dnd";
 import _ from "underscore";
 
-import BaseItemsTable from "metabase/collections/components/BaseItemsTable";
 import PinnedItemCard from "metabase/collections/components/PinnedItemCard";
 import BodyComponent from "metabase/components/BodyComponent";
+import { BaseItemsTable } from "metabase/components/ItemsTable/BaseItemsTable";
 
 // NOTE: our version of react-hot-loader doesn't play nice with react-dnd's DragLayer,
 // so we exclude files named `*DragLayer.jsx` in webpack.config.js
